+++ src/type_hash.c
@@ -9,7 +9,7 @@
 #include "rlite/page_btree.h"
 #include "rlite/util.h"
 
-static int rl_hash_create(rlite *db, long btree_page, rl_btree **btree)
+static int rl_hash_create(rlite *db, int64_t btree_page, rl_btree **btree)
 {
 	rl_btree *hash = NULL;
 
@@ -24,7 +24,7 @@ cleanup:
 	return retval;
 }
 
-static int rl_hash_read(rlite *db, long hash_page_number, rl_btree **btree)
+static int rl_hash_read(rlite *db, int64_t hash_page_number, rl_btree **btree)
 {
 	void *tmp;
 	int retval;
@@ -35,11 +35,11 @@ cleanup:
 	return retval;
 }
 
-static int rl_hash_get_objects(rlite *db, const unsigned char *key, long keylen, long *_hash_page_number, rl_btree **btree, int update_version, int create)
+static int rl_hash_get_objects(rlite *db, const unsigned char *key, int64_t keylen, int64_t *_hash_page_number, rl_btree **btree, int update_version, int create)
 {
-	long hash_page_number = 0, version = 0;
+	int64_t hash_page_number = 0, version = 0;
 	int retval;
-	unsigned long long expires = 0;
+	uint64_t expires = 0;
 	if (create) {
 		retval = rl_key_get_or_create(db, key, keylen, RL_TYPE_HASH, &hash_page_number, &version);
 		if (retval != RL_FOUND && retval != RL_NOT_FOUND) {
@@ -75,14 +75,14 @@ cleanup:
 	return retval;
 }
 
-int rl_hset(struct rlite *db, const unsigned char *key, long keylen, unsigned char *field, long fieldlen, unsigned char *data, long datalen, long *added, int update)
+int rl_hset(struct rlite *db, const unsigned char *key, int64_t keylen, unsigned char *field, int64_t fieldlen, unsigned char *data, int64_t datalen, int64_t *added, int update)
 {
 	int retval;
-	long hash_page_number;
+	int64_t hash_page_number;
 	rl_btree *hash;
 	unsigned char *digest = NULL;
 	rl_hashkey *hashkey = NULL;
-	long add = 1;
+	int64_t add = 1;
 	void *tmp;
 	RL_CALL(rl_hash_get_objects, RL_OK, db, key, keylen, &hash_page_number, &hash, 1, 1);
 
@@ -135,10 +135,10 @@ cleanup:
 	return retval;
 }
 
-int rl_hget(struct rlite *db, const unsigned char *key, long keylen, unsigned char *field, long fieldlen, unsigned char **data, long *datalen)
+int rl_hget(struct rlite *db, const unsigned char *key, int64_t keylen, unsigned char *field, int64_t fieldlen, unsigned char **data, int64_t *datalen)
 {
 	int retval;
-	long hash_page_number;
+	int64_t hash_page_number;
 	rl_btree *hash;
 	void *tmp;
 	unsigned char *digest = NULL;
@@ -160,17 +160,17 @@ cleanup:
 	return retval;
 }
 
-int rl_hmget(struct rlite *db, const unsigned char *key, long keylen, int fieldc, unsigned char **fields, long *fieldslen, unsigned char ***_data, long **_datalen)
+int rl_hmget(struct rlite *db, const unsigned char *key, int64_t keylen, int fieldc, unsigned char **fields, int64_t *fieldslen, unsigned char ***_data, int64_t **_datalen)
 {
 	int retval;
-	long hash_page_number;
+	int64_t hash_page_number;
 	rl_btree *hash;
 	void *tmp;
 	unsigned char *digest = NULL;
 	rl_hashkey *hashkey;
 
 	unsigned char **data = rl_malloc(sizeof(unsigned char *) * fieldc);
-	long *datalen = rl_malloc(sizeof(long) * fieldc);
+	int64_t *datalen = rl_malloc(sizeof(int64_t) * fieldc);
 	RL_CALL(rl_hash_get_objects, RL_OK, db, key, keylen, &hash_page_number, &hash, 0, 0);
 
 	RL_MALLOC(digest, sizeof(unsigned char) * 20);
@@ -203,10 +203,10 @@ cleanup:
 	return retval;
 }
 
-int rl_hmset(struct rlite *db, const unsigned char *key, long keylen, int fieldc, unsigned char **fields, long *fieldslen, unsigned char **datas, long *dataslen)
+int rl_hmset(struct rlite *db, const unsigned char *key, int64_t keylen, int fieldc, unsigned char **fields, int64_t *fieldslen, unsigned char **datas, int64_t *dataslen)
 {
 	int i, retval;
-	long hash_page_number;
+	int64_t hash_page_number;
 	rl_btree *hash;
 	unsigned char *digest = NULL;
 	rl_hashkey *hashkey = NULL;
@@ -251,10 +251,10 @@ cleanup:
 	return retval;
 }
 
-int rl_hexists(struct rlite *db, const unsigned char *key, long keylen, unsigned char *field, long fieldlen)
+int rl_hexists(struct rlite *db, const unsigned char *key, int64_t keylen, unsigned char *field, int64_t fieldlen)
 {
 	int retval;
-	long hash_page_number;
+	int64_t hash_page_number;
 	rl_btree *hash;
 	unsigned char *digest = NULL;
 	RL_CALL(rl_hash_get_objects, RL_OK, db, key, keylen, &hash_page_number, &hash, 0, 0);
@@ -268,15 +268,15 @@ cleanup:
 	return retval;
 }
 
-int rl_hdel(struct rlite *db, const unsigned char *key, long keylen, long fieldsc, unsigned char **fields, long *fieldslen, long *delcount)
+int rl_hdel(struct rlite *db, const unsigned char *key, int64_t keylen, int64_t fieldsc, unsigned char **fields, int64_t *fieldslen, int64_t *delcount)
 {
 	int retval;
-	long hash_page_number;
+	int64_t hash_page_number;
 	rl_btree *hash;
 	rl_hashkey *hashkey;
 	void *tmp;
-	long i;
-	long deleted = 0;
+	int64_t i;
+	int64_t deleted = 0;
 	int keydeleted = 0;
 	unsigned char *digest = NULL;
 	RL_CALL(rl_hash_get_objects, RL_OK, db, key, keylen, &hash_page_number, &hash, 1, 0);
@@ -312,7 +312,7 @@ cleanup:
 	return retval;
 }
 
-int rl_hgetall(struct rlite *db, rl_hash_iterator **iterator, const unsigned char *key, long keylen)
+int rl_hgetall(struct rlite *db, rl_hash_iterator **iterator, const unsigned char *key, int64_t keylen)
 {
 	int retval;
 	rl_btree *hash;
@@ -322,7 +322,7 @@ cleanup:
 	return retval;
 }
 
-int rl_hlen(struct rlite *db, const unsigned char *key, long keylen, long *len)
+int rl_hlen(struct rlite *db, const unsigned char *key, int64_t keylen, int64_t *len)
 {
 	int retval;
 	rl_btree *hash;
@@ -333,15 +333,15 @@ cleanup:
 	return retval;
 }
 
-int rl_hincrby(struct rlite *db, const unsigned char *key, long keylen, unsigned char *field, long fieldlen, long increment, long *newvalue)
+int rl_hincrby(struct rlite *db, const unsigned char *key, int64_t keylen, unsigned char *field, int64_t fieldlen, int64_t increment, int64_t *newvalue)
 {
 	int retval;
 	rl_btree *hash;
 	void *tmp;
 	unsigned char *digest = NULL, *data = NULL;
 	char *end;
-	long datalen, hash_page_number;
-	long long value;
+	int64_t datalen, hash_page_number;
+	int64_t value;
 	rl_hashkey *hashkey;
 
 	RL_CALL(rl_hash_get_objects, RL_OK, db, key, keylen, &hash_page_number, &hash, 1, 1);
@@ -380,7 +380,7 @@ int rl_hincrby(struct rlite *db, const unsigned char *key, long keylen, unsigned
 		rl_multi_string_delete(db, hashkey->value_page);
 
 		RL_MALLOC(data, sizeof(unsigned char) * MAX_LLONG_DIGITS);
-		datalen = snprintf((char *)data, MAX_LLONG_DIGITS, "%lld", value);
+		datalen = snprintf((char *)data, MAX_LLONG_DIGITS, "%" PRId64 "", value);
 		RL_CALL(rl_multi_string_set, RL_OK, db, &hashkey->value_page, data, datalen);
 
 		retval = rl_btree_update_element(db, hash, digest, hashkey);
@@ -397,7 +397,7 @@ int rl_hincrby(struct rlite *db, const unsigned char *key, long keylen, unsigned
 	}
 	else if (retval == RL_NOT_FOUND) {
 		RL_MALLOC(data, sizeof(unsigned char) * MAX_LLONG_DIGITS);
-		datalen = snprintf((char *)data, MAX_LLONG_DIGITS, "%ld", increment);
+		datalen = snprintf((char *)data, MAX_LLONG_DIGITS, "%" PRId64 "", increment);
 
 		RL_MALLOC(hashkey, sizeof(*hashkey));
 		RL_CALL(rl_multi_string_set, RL_OK, db, &hashkey->string_page, field, fieldlen);
@@ -417,14 +417,14 @@ cleanup:
 	return retval;
 }
 
-int rl_hincrbyfloat(struct rlite *db, const unsigned char *key, long keylen, unsigned char *field, long fieldlen, double increment, double *newvalue)
+int rl_hincrbyfloat(struct rlite *db, const unsigned char *key, int64_t keylen, unsigned char *field, int64_t fieldlen, double increment, double *newvalue)
 {
 	int retval;
 	rl_btree *hash;
 	void *tmp;
 	unsigned char *digest = NULL, *data = NULL;
 	char *end;
-	long dataalloc, datalen, hash_page_number;
+	int64_t dataalloc, datalen, hash_page_number;
 	double value;
 	rl_hashkey *hashkey;
 
@@ -497,7 +497,7 @@ cleanup:
 	return retval;
 }
 
-int rl_hash_iterator_next(rl_hash_iterator *iterator, long *fieldpage, unsigned char **field, long *fieldlen, long *memberpage, unsigned char **member, long *memberlen)
+int rl_hash_iterator_next(rl_hash_iterator *iterator, int64_t *fieldpage, unsigned char **field, int64_t *fieldlen, int64_t *memberpage, unsigned char **member, int64_t *memberlen)
 {
 	void *tmp;
 	rl_hashkey *hashkey = NULL;
@@ -546,7 +546,7 @@ int rl_hash_iterator_destroy(rl_hash_iterator *iterator)
 	return rl_btree_iterator_destroy(iterator);
 }
 
-int rl_hash_pages(struct rlite *db, long page, short *pages)
+int rl_hash_pages(struct rlite *db, int64_t page, short *pages)
 {
 	rl_btree *btree;
 	rl_btree_iterator *iterator = NULL;
@@ -583,7 +583,7 @@ cleanup:
 	return retval;
 }
 
-int rl_hash_delete(rlite *db, long value_page)
+int rl_hash_delete(rlite *db, int64_t value_page)
 {
 	rl_btree *hash;
 	rl_btree_iterator *iterator;
